pyproject.toml
[tool.poetry]
name = "skp-2-cw-6"
version = "0.1.0"
description = ""
authors = ["Stereo2025 <teneloks@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
django-crontab = "^0.7.1"
pytils = "^0.4.1"
redis = "^5.0.3"
ipython = "^8.23.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
